Move all the mock-signing code to test_utils::mock_signing
